Initial availability checker implementation.
In case of exception ping thread will be created.
This thread will send "ping" requests to the SAM Server.
If 3 request will be failed — availability checked decide that SAM server is not available.
In this case the SAM Agent will not send events to the SAM Server and continue sending just "ping" request. If "ping" pass then availability checker decide that SAM Server is available and SAM Agent continue sending SAM events to it.